Image distribution
Ran the function to find out the distribution and
mean values of the image data.

Collected the mean image shape and saved that into
a pkl-file. With this information I´ve came up
with a new hypothesis:
- Will the model learn better with the mean size
or with 224 x 224px?

Next step is to resize and save copies in a new
folder.