Had problem with committing
Big files wasn´t uploaded (.h5 or .keras)
so had to do a hard reset on the commits and
luckily I´ve committed right before the model
run. Now I remember that the model had a good
uprising curve, but doubled the dense layers in
last convolution layer. See if this works better.