Refined before model run no.2
Read a little bit more and I´ve made a mistake the
first time by copying every image to a new size,
that will 'ImageDataGenerator' do for me. So all
these copies will be deleted.

Since the first run was made with the mean value
of the images, this second run will be with
224 x 224 instead. Then see if the model learns
better with this size.