Soft git reset and model comprimizing
Noticed that I first deleted the wrong ipynb-file
after I´ve comprimized the model for this
ML pipeline. Also ran a test run with the model
with good result. But forgot to put in the save
function in the class. So tomorrow I will run my
version 1.

Also documented a lot on a piece of paper that I´m
going to add in my README.md tomorrow. Didn´t
wanna stop the model run to add this information.

Think I´m going to sort things out now with the
project! Hard start but now I´m on the right path!