Model run v1
Now the model has run with a good learning curve
so those are saved, but forgot to add a function
to save the model in my class. But images and
JSON-file are created. If I can´t tune the
hyperparameters to a better result this settings
will be ran again.