Feat: Color distribution
Added a function that iterates through all images
in the train map after the splitting of sets, to
show how similiar the colors are in the images.
tried first with all the images but it didn´t
look good, to much info to handle. Since it´s
training with the train set, it looked better to
present that information.